General refactor. Added class for settings -> configuration.dart Implemented basic window UI settings. Added report for transparent window. Added dartdoc.

+++ lib/core/widget/junction_widget_size.dart
@@ -6,7 +6,7 @@ enum JunctionWidgetSize {
 }
 
 extension JunctionWidgetSizeExtension on JunctionWidgetSize{
-  int? get size{
+  double? get size{
     switch(this){
       case JunctionWidgetSize.one:
         return 100;
